bench(kv): attention-softmax consumer results on Qwen2.5-1.5B + parquet-direct corpus download
KV serving bench (fp32 compute — Qwen fp16 attention logits overflow in the
BASELINE, NaN heads; measurement-care note kept in the script): the consumer
metric orders configs exactly as behavior does. hot512: attn-KL 0.016-0.054 ->
54% greedy-token agreement (first divergence ~step 26); hot128: 0.044-0.062 ->
37.5%; hot128/no-outliers: 0.047-0.065 -> 36.5%; all at ~3.8-4x nominal KV
compression. Headline for the Blindfold paper: hot512 and hot128 have
IDENTICAL nominal bytes/token — rate accounting cannot distinguish them; the
attention-KL consumer metric separates them and predicts the behavioral gap.

real_download.py rewritten to direct per-parquet hf_hub_download + pyarrow
(the datasets streaming client dies on long unauthenticated sessions and
cannot resume); resumable per corpus shard; 15M rows banked.

+++ benchmarks/bench_kv_serving.py
@@ -132,9 +132,13 @@ def main():
     args = ap.parse_args()
 
     tok = AutoTokenizer.from_pretrained(args.model)
+    # fp32, not fp16: Qwen attention logits overflow fp16 in eager capture
+    # (NaN heads in the *baseline*), and Volta-class GPUs have no native bf16.
+    # The KV cache under test is still stored/quantized by the adapter itself;
+    # only the compute dtype changes.
     model = AutoModelForCausalLM.from_pretrained(
         args.model,
-        dtype=torch.float16,
+        dtype=torch.float32,
         device_map="cuda:0",
         attn_implementation="eager",
     )
